Separate interfaces from Storage.
