+++ s2core/src/main/scala/org/apache/s2graph/core/storage/SKeyValue.scala
@@ -26,6 +26,7 @@ import org.hbase.async.KeyValue
 
 object SKeyValue {
   val EdgeCf = "e".getBytes("UTF-8")
+  val VertexCf = "v".getBytes("UTF-8")
   val Put = 1
   val Delete = 2
   val Increment = 3
