+++ s2core/src/main/scala/org/apache/s2graph/core/storage/serde/Deserializable.scala
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -17,14 +17,12 @@
  * under the License.
  */
 
-package org.apache.s2graph.core.storage
+package org.apache.s2graph.core.storage.serde
 
-import org.apache.hadoop.hbase.util.Bytes
-import org.apache.s2graph.core.types.{HBaseType, LabelWithDirection, SourceVertexId, VertexId}
+import org.apache.s2graph.core.types.{LabelWithDirection, VertexId}
 
 
 trait Deserializable[E] extends StorageDeserializable[E] {
-  import StorageDeserializable._
 
   type RowKeyRaw = (VertexId, LabelWithDirection, Byte, Boolean, Int)
